backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.3
matplotlib==3.9.2
networkx==2.8.8
openai==1.56.1
python-dotenv==1.0.1
gunicorn==21.2.0
chromadb==0.5.15
